test(builder): update spec for TaskBuilder
